backend/requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Database (install separately if needed)
# psycopg2-binary==2.9.9

# Environment and Configuration
python-decouple==3.8
django-environ==0.11.2

# CORS handling for API
django-cors-headers==4.3.1

# Authentication and Security
PyJWT==2.8.0
cryptography>=41.0.0

# Production Server
gunicorn==21.2.0

# Static Files (Production)
whitenoise==6.6.0

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
hypothesis==6.92.1
factory-boy==3.3.0

# Image processing
Pillow>=10.0.0

# API Documentation
drf-spectacular==0.27.0